skip term check
skip license header check

Conflicts: src/cpu/aarch64/matmul/acl_matmul.cpp

+++ src/gpu/intel/jit/gemm/generator/pieces/address_setup.cxx
@@ -366,7 +366,8 @@ void BLASKernelGenerator<hw>::setupAddr(Type T, const GRFRange &addr, const BO &
                 if (T.paddedSize() < widthAlign)
                     or_(1, addr[0].ud(2), addr[0].ud(2), widthAlign - 1);
             } else if (remW.isInvalid() && remH.isInvalid())
-                emov(1, addr[0].uq(1), uint64_t(bw * bcount * block.ebytes - 1) | (uint64_t(bh * block.ebytes - 1) << 32), strategy, state);
+                emov(1, addr[0].uq(1), (((uint64_t)bw * (uint64_t)bcount * block.ebytes - 1)
+                        | ((uint64_t)bh * block.ebytes - 1) << 32), strategy, state);
             else {
                 if (remW.isValid() && multiX > 1) stub();
                 remW.isValid() ? addScaled(1, addr[0].ud(2), -1, remW.uw(), T, state, true)
