+++ src/gpu/intel/jit/gemm/generator/pieces/matrix_access.cxx
@@ -516,7 +516,7 @@ void BLASKernelGenerator<hw>::loadLoadStoreDescriptors(bool load, bool store, Re
             descLoad.parts.responseLen = 0;
 
             int underlyingSIMD = std::max<int>(block.simdSize, maxScatteredSIMD(hw, astrategy) >> 1);
-            int log2GRFs = ilog2(underlyingSIMD * block.ebytes) - GRF::log2Bytes(hw);
+            int log2GRFs = ilog2((uint64_t)underlyingSIMD * block.ebytes) - GRF::log2Bytes(hw);
             int log2Components = int(block.splitComplex);
 
             if (channel) mov(1, t2, 0x1000 << log2Components);
